requirements.txt
langchain>=0.2.10
langchain-community>=0.2.10
faiss-cpu>=1.8.0
sentence-transformers>=2.2.2
numpy>=1.26
scikit-learn>=1.4
pypdf>=4.2.0
python-dotenv>=1.0.1
typer>=0.12.3
requests>=2.31.0
gradio>=4.43.0
tqdm>=4.66.0
